- fixes alignment issue on tree view
